moved the anvil mixin to it's own class, and renamed the block entity and it's type from Alternal anvil to just Anvil Block Entity
