+++ src/main/java/io/github/darealturtywurty/turtychemistry/common/TurtyTags.java
@@ -8,7 +8,7 @@ import net.minecraft.tags.TagKey;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.level.block.Block;
 
-public record TurtyTags() {
+public final class TurtyTags {
     public static final TagKey<Item> TURTY_ITEM_TAG_KEY = ItemTags.create(new ResourceLocation(TurtyChemistry.MODID));
     public static final TagKey<Block> TURTY_BLOCK_TAG_KEY = BlockTags.create(new ResourceLocation(TurtyChemistry.MODID));
 }
