+++ src/main/java/io/github/darealturtywurty/turtychemistry/common/item/BasicHammer.java
@@ -1,6 +1,6 @@
 package io.github.darealturtywurty.turtychemistry.common.item;
 
-import io.github.darealturtywurty.turtychemistry.common.block.entity.AlternateAnvilMechanicBlockEntity;
+import io.github.darealturtywurty.turtychemistry.common.block.entity.AnvilBlockEntity;
 import net.minecraft.world.InteractionResult;
 import net.minecraft.world.item.context.UseOnContext;
 import org.jetbrains.annotations.NotNull;
@@ -13,7 +13,7 @@ public final class BasicHammer extends ChemistryItem {
     @Override
     public @NotNull InteractionResult useOn(UseOnContext pContext) {
         if (pContext.getLevel().getBlockEntity(pContext.getClickedPos()) instanceof
-                AlternateAnvilMechanicBlockEntity alternateAnvil && pContext.getPlayer().isCrouching()) {
+                AnvilBlockEntity alternateAnvil && pContext.getPlayer().isCrouching()) {
             alternateAnvil.smithItem();
         }
         return super.useOn(pContext);
