+++ src/main/java/io/github/darealturtywurty/turtychemistry/core/init/BlockEntityInit.java
@@ -1,7 +1,7 @@
 package io.github.darealturtywurty.turtychemistry.core.init;
 
 import io.github.darealturtywurty.turtychemistry.TurtyChemistry;
-import io.github.darealturtywurty.turtychemistry.common.block.entity.AlternateAnvilMechanicBlockEntity;
+import io.github.darealturtywurty.turtychemistry.common.block.entity.AnvilBlockEntity;
 import io.github.darealturtywurty.turtychemistry.common.block.entity.ClayAlloyFurnaceBlockEntity;
 import io.github.darealturtywurty.turtylib.core.init.AbstractInit;
 import net.minecraft.world.level.block.Blocks;
@@ -16,7 +16,7 @@ public final class BlockEntityInit extends AbstractInit {
 
     public static final RegistryObject<BlockEntityType<ClayAlloyFurnaceBlockEntity>> CLAY_ALLOY_FURNACE = BLOCK_ENTITIES.register("clay_alloy_furnace",
             () -> BlockEntityType.Builder.of(ClayAlloyFurnaceBlockEntity::new, BlockInit.CLAY_ALLOY_FURNACE.get()).build(null));
-    public static final RegistryObject<BlockEntityType<AlternateAnvilMechanicBlockEntity>> ALTERNATE_ANVIL = BLOCK_ENTITIES.register("alternate_anvil",
-            () -> BlockEntityType.Builder.of(AlternateAnvilMechanicBlockEntity::new, Blocks.ANVIL,Blocks.CHIPPED_ANVIL,Blocks.DAMAGED_ANVIL).build(null));
+    public static final RegistryObject<BlockEntityType<AnvilBlockEntity>> ANVIL_BLOCK_ENTITY = BLOCK_ENTITIES.register("alternate_anvil",
+            () -> BlockEntityType.Builder.of(AnvilBlockEntity::new, Blocks.ANVIL,Blocks.CHIPPED_ANVIL,Blocks.DAMAGED_ANVIL).build(null));
 
 }
