+++ src/main/resources/turtychemistry.mixins.json
@@ -4,7 +4,7 @@
 	"compatibilityLevel": "JAVA_17",
 	"refmap": "turtychemistry.refmap.json",
 	"mixins": [
-		"MixinHooks$AnvilMixin"
+		"AnvilMixin"
 	],
 	"server": [],
 	"client": [],
